requirements.txt
millify
openpyxl
pandas
plotly
python-dotenv
scipy
streamlit
webcolors
